Feat/storage implement - uploadthing

+++ actions/exercises-action.ts
@@ -1,6 +1,7 @@
 'use server';
 
 import prisma from '@/lib/prisma';
+import { utapi } from '@/server/uploadthings';
 import type { ExerciseProps } from '@/types/data-types';
 import { revalidatePath } from 'next/cache';
 
@@ -36,10 +37,35 @@ export const createExercise = async (exerciseData: ExerciseProps): Promise<void>
   }
 };
 
+const deleteFileFromUploadThing = async (fileKey: string | string[]) => {
+  try {
+    await utapi.deleteFiles(fileKey);
+  } catch (error) {
+    throw new Error('Failed to delete file from UploadThing');
+  }
+};
+
 export const deleteExercise = async (exerciseId: string): Promise<void> => {
   if (!exerciseId) throw new Error('Missing ID for exercise deletion');
 
   try {
+    const exercise = await prisma.exercise.findUnique({
+      where: { id: exerciseId },
+    });
+
+    if (!exercise) throw new Error('Exercise not found');
+
+    try {
+      const url = new URL(exercise.imageUrl);
+      const fileKey = url.pathname.split('/').pop();
+      if (fileKey) {
+        await deleteFileFromUploadThing(fileKey);
+      }
+    } catch (error) {
+      console.error('Invalid image URL:', exercise.imageUrl, error);
+      throw new Error(error instanceof Error ? error.message : 'Unexpected error');
+    }
+
     await prisma.exercise.delete({
       where: {
         id: exerciseId,
@@ -59,6 +85,26 @@ export const updateExercise = async (exerciseData: ExerciseProps): Promise<void>
   if (!id) throw new Error('Missing ID for exercise update');
 
   try {
+    // Delete the old image from storage if the imageUrl is being updated
+    if (imageUrl) {
+      const existingExercise = await prisma.exercise.findUnique({
+        where: { id },
+      });
+
+      if (existingExercise?.imageUrl && existingExercise.imageUrl !== imageUrl) {
+        try {
+          const url = new URL(existingExercise.imageUrl);
+          const fileKey = url.pathname.split('/').pop();
+          if (fileKey) {
+            await deleteFileFromUploadThing(fileKey);
+          }
+        } catch (error) {
+          console.error('Invalid old image URL:', existingExercise.imageUrl, error);
+          throw new Error(error instanceof Error ? error.message : 'Unexpected error');
+        }
+      }
+    }
+
     await prisma.exercise.update({
       where: {
         id: id,
